add jwt authentication

+++ controllers/api/v1/posts_api.js
@@ -1,48 +1,51 @@
-const Post=require('../../../models/post')
-const Comment=require('../../../models/comment')
-module.exports.index= async function(req,res){
+const Post = require('../../../models/post');
+const Comment = require('../../../models/comment');
+module.exports.index = async function(req, res){
+
+
     let posts = await Post.find({})
-    .sort('-createdAt')
-    .populate('user')
-    .populate({
-        path: 'comments',
-        populate: {
-            path: 'user'
-        }
-    });
-    
-   return res.json(200,{
-        message:"list of posts",
-        posts:posts
+        .sort('-createdAt')
+        .populate('user')
+        .populate({
+            path: 'comments',
+            populate: {
+                path: 'user'
+            }
+        });
+
+    return res.json(200, {
+        message: "List of posts",
+        posts: posts
     })
 }
+
 
 module.exports.destroy = async function(req, res){
 
     try{
         let post = await Post.findById(req.params.id);
 
-        // if (post.user == req.user.id){
+        if (post.user == req.user.id){
             post.remove();
 
             await Comment.deleteMany({post: req.params.id});
 
 
-            // req.flash('success', 'Post and associated comments deleted!');
-             return res.status(200).json({
-                message:"success"
-            })
-
-            
-        // }else{
-        //     req.flash('error', 'You cannot delete this post!');
-        //     return res.redirect('back');
-        // }
+    
+            return res.json(200, {
+                message: "Post and associated comments deleted successfully!"
+            });
+        }else{
+            return res.json(401, {
+                message: "You cannot delete this post!"
+            });
+        }
 
     }catch(err){
-        return res.status(500).json({
-            message:"error"
-        })
+        console.log('********', err);
+        return res.json(500, {
+            message: "Internal Server Error"
+        });
     }
     
 }
